Add interface for db layer
