+++ src/application/core/domain/repositories/in-memory-receipt-repository.ts
@@ -2,41 +2,23 @@ import { Receipt } from '../../../../types/domain/receipt';
 import { GetPointsResponse } from '../../../../types/http/get-receipt-points';
 
 /**
- * A repository for storing receipts in memory.
- * Data is stored in a Map object for efficient key-value management.
- * Maps each receipt ID (string) to an object containing the receipt and its associated points.
+ * Interface for a receipt repository.
+ * It defines methods to save a receipt and retrieve receipt points.
  */
-export class InMemoryReceiptRepository {
+export interface IReceiptRepository {
   /**
-   * In-memory store for receipts.
-   */
-  private receipts: Map<string, { receipt: Receipt; points: number }> =
-    new Map();
-
-  /**
-   * Saves a receipt to the in-memory store.
-   * @param {string} id
-   * @param {Receipt} receipt
-   * @param {number} points
+   * Saves a receipt to the repository.
+   * @param {string} id - The receipt ID.
+   * @param {Receipt} receipt - The receipt data.
+   * @param {number} points - The points associated with the receipt.
    * @returns {string} The id of the stored receipt.
    */
-  save(id: string, receipt: Receipt, points: number): string {
-    this.receipts.set(id, { receipt, points });
-    return id;
-  }
+  save(id: string, receipt: Receipt, points: number): string;
 
   /**
-   * Finds a receipt by its ID.
-   * Retrieves the stored receipt and its associated points from the in-memory store.
-   * @param {string} id
-   * @returns {GetPointsResponse | null}
-   * An object containing the receipt and points if found, otherwise null.
+   * Finds a receipt by its ID and returns its associated points.
+   * @param {string} id - The receipt ID.
+   * @returns {GetPointsResponse | null} The points associated with the receipt, or null if not found.
    */
-  find(id: string): GetPointsResponse | null {
-    const savedReceipt = this.receipts.get(id);
-    if (!savedReceipt?.points) {
-      return null;
-    }
-    return { points: savedReceipt.points };
-  }
+  find(id: string): GetPointsResponse | null;
 }
